refactor(protocol): merge the opportunity presentation cluster into one module
