+++ packages/protocol/src/opportunities/application/index.ts
@@ -40,11 +40,11 @@ export * from "./opportunity.persistence-admission.js";
 export * from "./opportunity.update-admission.js";
 export * from "./opportunity.newborn-stamping.js";
 export * from "./opportunity.evaluator.js";
-export * from "./opportunity.presenter.js";
+export * from "./opportunity.presentation.js";
 export * from "./opportunity.introducer.js";
 export * from "./opportunity.existing-negotiation.js";
 export * from "./opportunity.graph.js";
-export * from "./opportunity.card-presentation.js";
+export * from "./opportunity.presentation.js";
 export * from "./opportunity.discovery-negotiation-summary.js";
 export * from "./negotiation-context.loader.js";
 export * from "./opportunity.feed-selection.js";
