+++ packages/protocol/src/opportunities/domain/index.ts
@@ -31,9 +31,9 @@ export * from "./opportunity.utils.js";
 export * from "./opportunity.actor.js";
 export * from "./opportunity.evidence.js";
 export * from "../../shared/utils/claim-safety.js";
-export * from "./opportunity.presentation.js";
-export * from "./opportunity.safe-presentation.js";
-export * from "./opportunity.presentation-cache.js";
+export * from "../application/opportunity.presentation.js";
+export * from "../application/opportunity.presentation.js";
+export * from "../application/opportunity.presentation.js";
 export * from "./negotiation-summary.builder.js";
 
 // ── Subdirectory domain exports ───────────────────────────────────────────────
