+++ packages/protocol/src/opportunities/radar/radar.graph.ts
@@ -16,12 +16,12 @@ import { StateGraph, START, END } from '@langchain/langgraph';
 import type { RadarGraphDatabase, OpportunityStatus } from '../../shared/interfaces/database.interface.js';
 import type { OpportunityCache } from '../../shared/interfaces/cache.interface.js';
 import { RadarGraphState, type RadarCardItem, type RadarResponseItem } from './radar.state.js';
-import { OpportunityPresenter, gatherPresenterContext, type PresenterDatabase } from '../application/opportunity.presenter.js';
+import { OpportunityPresenter, gatherPresenterContext, type PresenterDatabase } from '../application/opportunity.presentation.js';
 import { loadNegotiationContext } from '../application/negotiation-context.loader.js';
 import { canUserSeeOpportunity, isActionableForViewer, selectByComposition } from '../domain/opportunity.utils.js';
 import { getPrimaryActionLabel, SECONDARY_ACTION_LABEL } from '../domain/opportunity.labels.js';
-import { safeFallbackSummary } from '../domain/opportunity.safe-presentation.js';
-import { buildRadarCardPresentationCacheKey } from '../domain/opportunity.presentation-cache.js';
+import { safeFallbackSummary } from '../application/opportunity.presentation.js';
+import { buildRadarCardPresentationCacheKey } from '../application/opportunity.presentation.js';
 import type { DebugMetaAgent } from '../../agents/index.js';
 import { protocolLogger } from '../../shared/observability/protocol.logger.js';
 import { timed } from '../../shared/observability/performance.js';
