+++ packages/protocol/src/opportunities/tests/opportunity.card-text.spec.ts
@@ -3,7 +3,7 @@ config({ path: '.env.test', override: true });
 
 import { describe, expect, it } from "bun:test";
 
-import { viewerCentricCardSummary, narratorRemarkFromReasoning } from "../domain/opportunity.presentation.js";
+import { narratorRemarkFromReasoning, viewerCentricCardSummary } from "../application/opportunity.presentation.js";
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // viewerCentricCardSummary
