+++ packages/protocol/src/opportunities/tests/opportunity.sanitize.spec.ts
@@ -2,7 +2,7 @@ import { config } from "dotenv";
 config({ path: ".env.test", override: true });
 
 import { describe, expect, it } from "bun:test";
-import { stripIntroducerMentions } from "../domain/opportunity.presentation.js";
+import { stripIntroducerMentions } from "../application/opportunity.presentation.js";
 
 describe("stripIntroducerMentions", () => {
   it("removes introducer mention at start of sentence", () => {
